Main site Beginging
The main site code has begun ! I have created the CSS and Java files and had to remove the Typo-Formal font as the file format was not compatible. I have also added the backing image for the site but am likely to tone the yellow down and sort the image out so it looks cleaner as the yellow has a gap along the right-hand-side.