Button Working Kind of
The button which is going to create a pop out window for users to log in is working. But only when the HTML, Java and CSS is all within the the HTML file. When putting Java in the java sheet it breaks down, the same happens with the CSS. Also the button is badly placed along with the pop up menu but that is of no concern A.T.M. (At The Moment)