.Json tests
This consists of sorting out a few issues with sizing of elements on a different screen size *NOTE* this must be tested to see if works on many a screen sizes.*. Then console.logs added to determine where the Java code goes in relation to the .Json code, this has show that the code does NOT enter line 87 and skips down to line 100. although the entire code block is skipped and run after line 103.