Log in Detection
The log in system is getting function. When details provided in the log in box they are stored in Username and Password variables respectively, then being outputted (Just for testing purposes). Then a few .Json ideas created and read internally in the java file rather than a separate .Json file. More testing required.